refactor: dataframe and query_engine
